+++ src/Serv/Server/Core/Config.hs
@@ -1,0 +1,7 @@
+module Serv.Server.Core.Config
+    ( module Serv.Server.Core.Config.Types
+    , module Serv.Server.Core.Config.ConfigLoader
+    ) where
+
+import           Serv.Server.Core.Config.ConfigLoader
+import           Serv.Server.Core.Config.Types