+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+0.1.0.3
+-------
+Server Management:
+  - secured /metrics
+  - configuration yaml supports server, management, logging configuration
+      server:
+        name:
+        port:
+      management:
+        port:
+        userName:
+        userPassword:
+      log:
+        appenders:
+        - Console
+        request: Console
+
 0.1.0.2
 -------
 Logging improvements: configure request log appender - None, Console, File (req.log)